Made destructor of module virtual and overrided it modules.
This was a serious issue, as it was causing memory leak of some members that are created on the stack but has internal dynamic memory allocation in heap. Such as, adding std::vector<float> to ModuleEditor, even leaving it untouched, was causing memory leak. This way, when the application gets deleted all the modules will be deleted too properly.
For explicitness, all the destructors are marked with override keyword.